[Common] Merge the support/casting.h into common/types.h.
