+++ include/runtime/instance/memory.h
@@ -14,8 +14,8 @@
 #include "common/ast/type.h"
 #include "common/errcode.h"
 #include "common/log.h"
+#include "common/types.h"
 #include "common/value.h"
-#include "support/casting.h"
 #include "support/span.h"
 
 #include <algorithm>
