+++ lib/interpreter/engine/engine.cpp
@@ -4,7 +4,6 @@
 #include "common/statistics.h"
 #include "common/value.h"
 #include "interpreter/interpreter.h"
-#include "support/casting.h"
 #include "support/measure.h"
 #include "support/time.h"
